CRAYSAT-1031: Add removal from product catalog
This commit adds support for removing entries from the product catalog,
by running the `catalog_delete.py` script using subprocess. It is
assumed that the `catalog_delete.py` will be installed and available
on the $PATH.

Test Description: I built the sat-install-utility image and ran an
uninstall, verifying that the version being uninstalled was removed
from the product catalog.