CRAYSAT-1033: Add support for multiple images per product
This commit adds support for product catalog entries that are structured
as follows:

  component_versions:
    docker:
      image_name: image_version
      ...
    rpm:
      package-name: package_version
      ...

The logic for determining shared images is modified slightly, such that
a product being uninstalled will have all of its Docker images removed,
but only the ones that are not referenced by any other product.

Unit tests are updated accordingly.

Add change log entry.

Test Description: I installed a newer version of sat-product-stream
on Thanos that incldues a product catalog entry for each Docker image,
and attempted to uninstall it. I was able to uninstall the CFS and
install utility images, but not the SAT image, as this was shared
by the previous install.