CRAYSAT-1031: Import catalog_delete.py from Docker image
This commit changes the method of including deletion from the
cray-product-catalog Kubernetes config map, from using a Python
dependency to using a multi-stage Docker build to copy in the script.

This commit also updates the revision of cray-product-catalog-update
used, from a personal fork to a published cray-product-catalog-update
image, although it is still in an 'unstable' repository.

* Remove cray-product-catalog from the requirements files.
* Modify Dockerfile to pull in catalog_delete.py

Test Description: ran the following command and checked that <version>
was removed from the product catalog.

  \# podman run --rm -v $kube_config_file:$HOME/.kube/config:ro \
      -v $cert_src_dir:$cert_target_dir:ro <thisImage> uninstall sat \
      <version>