CRAYSAT-1051: Remove old repos from group
This commit changes the behavior of the 'activate' action.

Previously, activating a particular product version would set that
version's hosted-type repository as the first member in a list of
members of the product's group-type package repository in Nexus.

Now, activating a particular product version sets that version's
hosted-type repository as the only member in the group-type repository.

Add change log entry for this change.

Test Description:
I installed multiple versions of the SAT product stream, and used
the 'activate' action to switch between them. Between runs, I checked
the Nexus repositories.