CRAYSAT-1167: Use common library for CLI parsing
This removes the create_parser function and instead imports it
from shasta_install_utility_common, to keep common code in a
common location.

This also updates the version of shasta-install-utility-common
in the pinned requirements files to one that provides the new
parser module.

Test Description:
I installed the branch of shasta_install_utility_common along
with this branch in a virtualenv, and ran
`sat-install-utility --help`. I also ran unit tests.