CRAYSAT-1187: Integrate cfs-config-util library
This change adds a new depenedency on the the `cfs-config-util` library.
This allows `sat-install-utility` to leverage the CFS automation
functionality of this library to automatically update configurations
when downgrading/activating or uninstalling the SAT product stream.

Test Description:
Run `sat-install-utility` via prodmgr on thanos to ensure CFS
configuration is updated properly.