CRAYSAT-1311: Add missing dependencies to lock files
I noticed that a few dependencies were missing from the
requirements.lock.txt and requirements-dev.lock.txt. This commit adds
them to the list.

I generated this file by building the container image without this
change and then running `pip freeze` inside the container.

Test Description: I built the docker image again and checked that the
output of `pip freeze` was the same.