CRAYSAT-1750: Upgrade dependencies to allow newer PyYAMl
Upgrade dependencies to allow PyYAML 6.0.1. This version of PyYAML fixes
a build problem which occurs when building PyYAML in Alpine Linux
containers.

Also remove PyYAML from the top-level dependencies of this package
because it is actually not a direct dependency, but rather introduced
through the dependencies on cfs-config-util and
shasta-install-utility-common.

Test Description:
Unit tests pass. Will test the build from this branch on mug by running
it with the `activate` argument, which is the only argument that
`prodmgr` uses with this container moving forward and is the argument
for which code was most impacted.